pyproject.toml
# Copyright (c) 2022 Joseph Hale
# 
# This Source Code Form is subject to the terms of the Mozilla Public
# License, v. 2.0. If a copy of the MPL was not distributed with this
# file, You can obtain one at http://mozilla.org/MPL/2.0/.

[tool.poetry]
name = "git-authorship"
version = "0.1.0"
description = ""
authors = ["Joseph Hale <me@jhale.dev>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.27"
plotly = "^5.10.0"
gitpython = {git = "https://github.com/gitpython-developers/GitPython", rev = "bec6157"}

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
mypy = "^0.971"
pre-commit = "^2.20.0"
pytest = "^7.1.3"
liccheck = "^0.7.2"

[tool.liccheck]
authorized_licenses = [
    "MIT",
    "Apache-2.0",
    "Apache Software",
    "Apache License, Version 2.0",
    "Apache License 2.0",
    "BSD",
    "BSD-2-Clause",
    "BSD-3-Clause",
    "LGPL-2.1",
    "LGPL-3.0",
    "MPL-2.0",
    "Mozilla Public License 2.0 (MPL 2.0)",
    "Unlicense"
]
unautorized_licenses = [
    "GPL",
    "GNU General Public License v2 (GPLv2)"
]

[tool.liccheck.authorized_packages]
# None right now.

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
